Split SDNQ into multiple files and linting
